app/requirements.txt
retry
pyinstaller
ttkthemes
matplotlib
azure-mgmt-datalake-store
azure-mgmt-resource
azure-datalake-store
fredapi
python-bcb
numpy
pandas
openpyxl